refactor(analytics): AnalyticsPage 정의를 도메인 모듈로 이동 (기계적)
1,691줄 단일 파일에서 정의를 이름 그대로 9개 파일로 옮겼다 —
types 143(타입 17종) · utils 190(헬퍼 11종) · EvalDetailView 164 ·
constants 63(상수 7종) · api 59(함수 5종) · ActivityHeatmapChart 79 ·
CostComparisonCard 43 · MetricCard 39 · ChartCard 35. 메인은 958줄.

**958 은 한도 초과이며 그것이 의도다.** 계획서가 3a(기계적 이동)와
3b(섹션 추출)를 가른 이유는 B3 의 Task 4/5·7/8 과 같다 — 이름 그대로
옮기는 작업과 새 컴포넌트 경계·props 를 만드는 작업을 한 커밋에 섞으면
리뷰가 둘을 구별할 수 없다. 800 이내 진입은 3b 가 한다.

계획서 인벤토리의 오차 하나를 정정했다. 64-311 을 "타입 17종 + 상수 6종"
으로 묶었으나 실제로는 API 레이어(fetchDashboard 외 4종)가 섞여 있었고,
상수 2종(effortColors·strategyIcons)이 API 함수들 **사이에**(284-296) 끼어
있었다. 섹션 배너 기준으로 잘랐다면 상수가 api.ts 로 휩쓸려 갔을 것이다 —
정의 단위 경계로 자른 이유가 이것이다. Task 1(NotificationRuleEditor)에서
같은 오차가 났고 같은 처방(api 분리)을 썼다.

배치는 components/analytics/ 다. Task 2 의 "공유 디렉토리면 중첩" 규칙은
여기서 발동하지 않는다 — components/analytics/ 는 컴포넌트 1개 + index.ts
뿐이라 Task 1 의 notifications/ 와 같은 상황이고, 이 레포의 지배적 관용도
"페이지의 부품은 components/<도메인>/ 에 산다"이다(project-configs 의 탭
9종, workflows, organizations, monitor 가 전부 그 형태). pages/ 에는 하위
디렉토리가 하나도 없어 pages/analytics/ 쪽이 오히려 신설 관례가 된다.

utils 를 .ts 가 아니라 .tsx 로 둔 것은 renderAosModelSourceBadge 가 JSX 를
반환하기 때문이다. 컴포넌트로 승격하거나 이름을 바꾸는 것은 판단이라
3a 의 범위가 아니다.

index.ts 는 손대지 않았다(ProjectMultiSelect 재노출만 유지). 9개 전부
소비자가 AnalyticsPage 하나뿐인 내부 전용이다.

메인 파일이 제자리에 남으므로 routes.tsx:59 의 named import 위험이 없다.
계획서 경고대로 그 줄은 `m.AnalyticsPage as React.ComponentType<...>` 라
as 캐스트가 이름 유실을 가리지만, 파일을 옮기지 않았으므로 해당 없음.
named + default export 둘 다 원본 그대로 유지했다.

게이트: tsc 0(첫 시도, import 추정 오차 0) · ESLint 0 · vitest 205 파일
4,365 테스트 전부 통과(Task 2 와 동일 수치 — collection 유실 없음) ·
build 성공. 추가로 dist 에 AnalyticsPage 청크(50.64 kB)가 여전히 분리돼
있어 lazy 경계가 보존됐음을 확인했다 — 부품이 다른 곳에서 정적 import
되면 청크가 합쳐지고 INEFFECTIVE_DYNAMIC_IMPORT 경고가 나는데, 그 경고
대상은 기존 RegisterPage·InvitationAcceptPage 뿐이다.
