+++ src/dashboard/src/components/analytics/MetricCard.tsx
@@ -1,0 +1,39 @@
+/**
+ * 단일 지표 카드 — AnalyticsPage 내부 전용.
+ */
+
+import { Zap } from 'lucide-react'
+import { cn } from '@/lib/utils'
+
+interface MetricCardProps {
+  title: string
+  value: string | number
+  icon: typeof Zap
+  color: 'blue' | 'green' | 'red' | 'amber' | 'purple'
+  subtitle?: string
+}
+
+export function MetricCard({ title, value, icon: Icon, color, subtitle }: MetricCardProps) {
+  const colorClasses = {
+    blue: 'bg-blue-100 text-blue-600 dark:bg-blue-900/30 dark:text-blue-400',
+    green: 'bg-green-100 text-green-600 dark:bg-green-900/30 dark:text-green-400',
+    red: 'bg-red-100 text-red-600 dark:bg-red-900/30 dark:text-red-400',
+    amber: 'bg-amber-100 text-amber-600 dark:bg-amber-900/30 dark:text-amber-400',
+    purple: 'bg-purple-100 text-purple-600 dark:bg-purple-900/30 dark:text-purple-400',
+  }
+
+  return (
+    <div className="bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 p-4">
+      <div className="flex items-center justify-between mb-2">
+        <span className="text-sm text-gray-500 dark:text-gray-400">{title}</span>
+        <div className={cn('p-2 rounded-lg', colorClasses[color])}>
+          <Icon className="w-4 h-4" />
+        </div>
+      </div>
+      <div className="text-2xl font-bold text-gray-900 dark:text-white">{value}</div>
+      {subtitle && (
+        <div className="text-xs text-gray-500 dark:text-gray-400 mt-1">{subtitle}</div>
+      )}
+    </div>
+  )
+}
