+++ src/dashboard/src/pages/AnalyticsPage.tsx
@@ -40,274 +40,46 @@ import {
   MessageSquare,
   ChevronDown,
   ChevronRight,
-  ArrowRight,
-  GitBranch,
   Loader2,
-  FileText,
 } from 'lucide-react'
 import { cn } from '../lib/utils'
-import { computeHeatmapAlpha } from '../lib/heatmap'
 import { useProjectsStore, Project } from '../stores/projects'
-import {
-  useExternalUsageStore,
-  type ExternalUsageSummaryResponse,
-} from '../stores/externalUsage'
+import { useExternalUsageStore } from '../stores/externalUsage'
 import { useAuthStore } from '../stores/auth'
 import { ProjectMultiSelect } from '../components/analytics/ProjectMultiSelect'
 import type { TaskAnalysisHistory } from '../stores/agents'
-import { apiClient } from '@/services/apiClient'
-
-// ─────────────────────────────────────────────────────────────
-// Types
-// ─────────────────────────────────────────────────────────────
-
-type TimeRange = '1h' | '24h' | '7d' | '30d' | 'all'
-
-interface OverviewMetrics {
-  total_sessions: number
-  active_sessions: number
-  total_tasks: number
-  completed_tasks: number
-  failed_tasks: number
-  pending_tasks: number
-  success_rate: number
-  total_tokens: number
-  total_cost: number
-  avg_task_duration_ms: number
-  approvals_pending: number
-  approvals_granted: number
-  approvals_denied: number
-}
-
-interface TrendDataPoint {
-  timestamp: string
-  value: number | null
-  label: string
-}
-
-interface MultiTrendData {
-  time_range: TimeRange
-  tasks: TrendDataPoint[]
-  success_rate: TrendDataPoint[]
-  costs: TrendDataPoint[]
-  tokens: TrendDataPoint[]
-}
-
-interface AgentPerformance {
-  agent_id: string
-  agent_name: string
-  category: string
-  total_tasks: number
-  completed_tasks: number
-  failed_tasks: number
-  success_rate: number
-  avg_duration_ms: number
-  total_tokens: number
-  total_cost: number
-}
-
-interface CostBreakdown {
-  category: string
-  value: string
-  provider?: string | null
-  cost: number
-  tokens: number
-  percentage: number
-}
-
-interface CostAnalytics {
-  time_range: TimeRange
-  total_cost: number
-  total_tokens: number
-  avg_cost_per_task: number
-  by_agent: CostBreakdown[]
-  by_model: CostBreakdown[]
-  projected_monthly: number
-}
-
-interface HeatmapCell {
-  day: number
-  hour: number
-  value: number
-}
-
-interface ActivityHeatmap {
-  cells: HeatmapCell[]
-  max_value: number
-  time_range: TimeRange
-}
-
-interface AnalyticsDashboard {
-  overview: OverviewMetrics
-  trends: MultiTrendData
-  agents: { agents: AgentPerformance[]; time_range: TimeRange }
-  costs: CostAnalytics
-  activity: ActivityHeatmap
-}
-
-interface ModelTokenBreakdown {
-  model: string
-  provider: string
-  providerLabel: string
-  tokens: number
-  cost: number
-  percentage: number
-  color: string
-}
-
-// Multi-Project Comparison Types
-interface ProjectTrendSeries {
-  project_id: string
-  project_name: string
-  color: string
-  data: TrendDataPoint[]
-}
-
-interface MultiProjectTrendsResponse {
-  metric: string
-  period: TimeRange
-  series: ProjectTrendSeries[]
-}
-
-type CompareMetric = 'tasks' | 'tokens' | 'cost' | 'success_rate'
-
-interface AgentEvalStats {
-  agent_id: string
-  avg_rating: number
-  accuracy_rate: number
-  speed_satisfaction_rate: number
-  total_count: number
-}
-
-interface TaskEvalStats {
-  avg_rating: number
-  accuracy_rate: number
-  speed_satisfaction_rate: number
-  total_count: number
-  by_agent: AgentEvalStats[]
-}
-
-interface TaskEvaluation {
-  id: string
-  session_id: string
-  task_id: string
-  rating: number
-  result_accuracy: boolean
-  speed_satisfaction: boolean
-  comment: string | null
-  agent_id: string | null
-  created_at: string
-}
-
-// ─────────────────────────────────────────────────────────────
-// Constants
-// ─────────────────────────────────────────────────────────────
-
-const TIME_RANGES: { label: string; value: TimeRange }[] = [
-  { label: '1 Hour', value: '1h' },
-  { label: '24 Hours', value: '24h' },
-  { label: '7 Days', value: '7d' },
-  { label: '30 Days', value: '30d' },
-  { label: 'All Time', value: 'all' },
-]
-
-const CHART_COLORS = [
-  '#3B82F6', // blue
-  '#10B981', // green
-  '#F59E0B', // amber
-  '#EF4444', // red
-  '#8B5CF6', // purple
-  '#EC4899', // pink
-  '#06B6D4', // cyan
-]
-
-const PROVIDER_COLORS: Record<string, string> = {
-  codex_cli: '#8B5CF6',
-  claude_cli: '#F59E0B',
-  anthropic: '#F97316',
-  openai: '#10B981',
-  google: '#3B82F6',
-  github_copilot: '#111827',
-  ollama: '#64748B',
-  unknown: '#6B7280',
-}
-
-const PROVIDER_LABELS: Record<string, string> = {
-  codex_cli: 'Codex CLI',
-  claude_cli: 'Claude CLI',
-  anthropic: 'Anthropic Claude',
-  openai: 'OpenAI',
-  google: 'Google Gemini',
-  github_copilot: 'GitHub Copilot',
-  ollama: 'Ollama',
-  unknown: 'Unknown',
-}
-
-const DAY_LABELS = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat']
-
-// ─────────────────────────────────────────────────────────────
-// API
-// ─────────────────────────────────────────────────────────────
-
-async function fetchDashboard(timeRange: TimeRange, projectId?: string): Promise<AnalyticsDashboard> {
-  const params = new URLSearchParams({ time_range: timeRange })
-  if (projectId) params.append('project_id', projectId)
-  return apiClient.get<AnalyticsDashboard>(`/api/analytics/dashboard?${params}`)
-}
-
-async function fetchTaskEvalStats(projectId?: string): Promise<TaskEvalStats> {
-  const params = new URLSearchParams()
-  if (projectId) params.set('project_id', projectId)
-  const qs = params.toString()
-  return apiClient.get<TaskEvalStats>(`/api/feedback/task-evaluation/stats${qs ? `?${qs}` : ''}`)
-}
-
-async function fetchTaskEvalList(
-  agentId?: string,
-  limit = 50,
-  projectId?: string,
-): Promise<TaskEvaluation[]> {
-  const params = new URLSearchParams({ limit: String(limit) })
-  if (agentId) params.set('agent_id', agentId)
-  if (projectId) params.set('project_id', projectId)
-  return apiClient.get<TaskEvaluation[]>(`/api/feedback/task-evaluation/list?${params}`)
-}
-
-async function fetchAnalysisDetail(analysisId: string): Promise<TaskAnalysisHistory | null> {
-  try {
-    return await apiClient.get<TaskAnalysisHistory>(`/api/agents/orchestrate/analyses/${analysisId}`)
-  } catch {
-    return null
-  }
-}
-
-// Effort level colors (matches TaskAnalyzer)
-const effortColors: Record<string, string> = {
-  quick: 'bg-green-100 text-green-700 dark:bg-green-900/30 dark:text-green-400',
-  medium: 'bg-yellow-100 text-yellow-700 dark:bg-yellow-900/30 dark:text-yellow-400',
-  thorough: 'bg-red-100 text-red-700 dark:bg-red-900/30 dark:text-red-400',
-}
-
-// Strategy icons (matches TaskAnalyzer)
-const strategyIcons: Record<string, typeof GitBranch> = {
-  sequential: ArrowRight,
-  parallel: Zap,
-  mixed: GitBranch,
-}
-
-async function fetchMultiProjectTrends(
-  projectIds: string[],
-  metric: CompareMetric,
-  timeRange: TimeRange
-): Promise<MultiProjectTrendsResponse> {
-  const params = new URLSearchParams({ metric, time_range: timeRange })
-  projectIds.forEach((id) => params.append('project_ids', id))
-  return apiClient.get<MultiProjectTrendsResponse>(`/api/analytics/trends/compare?${params}`)
-}
-
-// ─────────────────────────────────────────────────────────────
-// Component
-// ─────────────────────────────────────────────────────────────
+import {
+  fetchAnalysisDetail,
+  fetchDashboard,
+  fetchMultiProjectTrends,
+  fetchTaskEvalList,
+  fetchTaskEvalStats,
+} from '../components/analytics/api'
+import { CHART_COLORS, TIME_RANGES } from '../components/analytics/constants'
+import type {
+  AnalyticsDashboard,
+  CompareMetric,
+  MultiProjectTrendsResponse,
+  TaskEvalStats,
+  TaskEvaluation,
+  TimeRange,
+} from '../components/analytics/types'
+import {
+  buildModelTokenBreakdown,
+  filterAttributedModelPerformance,
+  formatDuration,
+  formatNumber,
+  formatTokenCount,
+  formatTrendData,
+  renderAosModelSourceBadge,
+  transformMultiSeriesData,
+  truncateModelLabel,
+} from '../components/analytics/utils'
+import { ActivityHeatmapChart } from '../components/analytics/ActivityHeatmapChart'
+import { ChartCard } from '../components/analytics/ChartCard'
+import { CostComparisonCard } from '../components/analytics/CostComparisonCard'
+import { EvalDetailView } from '../components/analytics/EvalDetailView'
+import { MetricCard } from '../components/analytics/MetricCard'
 
 export function AnalyticsPage() {
   const [timeRange, setTimeRange] = useState<TimeRange>('7d')
@@ -1181,511 +953,6 @@ export function AnalyticsPage() {
      
     </div>
   )
-}
-
-// ─────────────────────────────────────────────────────────────
-// Sub-components
-// ─────────────────────────────────────────────────────────────
-
-interface MetricCardProps {
-  title: string
-  value: string | number
-  icon: typeof Zap
-  color: 'blue' | 'green' | 'red' | 'amber' | 'purple'
-  subtitle?: string
-}
-
-function MetricCard({ title, value, icon: Icon, color, subtitle }: MetricCardProps) {
-  const colorClasses = {
-    blue: 'bg-blue-100 text-blue-600 dark:bg-blue-900/30 dark:text-blue-400',
-    green: 'bg-green-100 text-green-600 dark:bg-green-900/30 dark:text-green-400',
-    red: 'bg-red-100 text-red-600 dark:bg-red-900/30 dark:text-red-400',
-    amber: 'bg-amber-100 text-amber-600 dark:bg-amber-900/30 dark:text-amber-400',
-    purple: 'bg-purple-100 text-purple-600 dark:bg-purple-900/30 dark:text-purple-400',
-  }
-
-  return (
-    <div className="bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 p-4">
-      <div className="flex items-center justify-between mb-2">
-        <span className="text-sm text-gray-500 dark:text-gray-400">{title}</span>
-        <div className={cn('p-2 rounded-lg', colorClasses[color])}>
-          <Icon className="w-4 h-4" />
-        </div>
-      </div>
-      <div className="text-2xl font-bold text-gray-900 dark:text-white">{value}</div>
-      {subtitle && (
-        <div className="text-xs text-gray-500 dark:text-gray-400 mt-1">{subtitle}</div>
-      )}
-    </div>
-  )
-}
-
-interface CostComparisonCardProps {
-  estimatedCost: number
-  estimatedTokens: number
-  actualCost: number | null
-}
-
-function CostComparisonCard({ estimatedCost, estimatedTokens, actualCost }: CostComparisonCardProps) {
-  const formatCostValue = (cost: number): string => {
-    if (cost === 0) return 'FREE'
-    if (cost < 0.01) return `$${cost.toFixed(4)}`
-    return `$${cost.toFixed(2)}`
-  }
-
-  return (
-    <div className="bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 p-4">
-      <div className="flex items-center justify-between mb-2">
-        <span className="text-sm text-gray-500 dark:text-gray-400">Total Cost</span>
-        <div className={cn('p-2 rounded-lg', 'bg-amber-100 text-amber-600 dark:bg-amber-900/30 dark:text-amber-400')}>
-          <DollarSign className="w-4 h-4" />
-        </div>
-      </div>
-      <div className="text-2xl font-bold text-gray-900 dark:text-white">
-        {formatCostValue(estimatedCost)}
-      </div>
-      <div className="text-xs text-gray-500 dark:text-gray-400 mt-1">
-        {formatNumber(estimatedTokens)} tokens
-        {actualCost !== null && (
-          <span className="ml-1 text-amber-600 dark:text-amber-400">
-            (실제 과금: {formatCostValue(actualCost)})
-          </span>
-        )}
-      </div>
-    </div>
-  )
-}
-
-interface ChartCardProps {
-  title: string
-  icon: typeof BarChart3
-  children: React.ReactNode
-  className?: string
-  headerExtra?: React.ReactNode
-}
-
-function ChartCard({ title, icon: Icon, children, className, headerExtra }: ChartCardProps) {
-  return (
-    <div
-      className={cn(
-        'bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 p-8',
-        className
-      )}
-    >
-      <div className="flex items-center justify-between mb-4 gap-2">
-        <h3 className="text-sm font-medium text-gray-900 dark:text-white flex items-center gap-2">
-          <Icon className="w-4 h-4 text-gray-400" />
-          {title}
-        </h3>
-        {headerExtra}
-      </div>
-      {children}
-    </div>
-  )
-}
-
-function EvalDetailView({ detail }: { detail: TaskAnalysisHistory }) {
-  const analysis = detail.analysis
-
-  return (
-    <div className="bg-gray-50 dark:bg-gray-900 rounded-lg p-4 space-y-4 mt-2">
-      {/* Task Input */}
-      <div className="flex items-start gap-2">
-        <FileText className="w-4 h-4 text-gray-400 mt-0.5 flex-shrink-0" />
-        <div>
-          <p className="text-xs font-medium text-gray-500 dark:text-gray-400 mb-1">태스크 입력</p>
-          <p className="text-sm text-gray-700 dark:text-gray-300 whitespace-pre-wrap line-clamp-3">
-            {detail.task_input}
-          </p>
-        </div>
-      </div>
-
-      {/* Analysis Summary 4-grid */}
-      {analysis && (
-        <>
-          <div className="grid grid-cols-2 md:grid-cols-4 gap-3">
-            <div className="bg-white dark:bg-gray-800 rounded-lg p-3 border border-gray-200 dark:border-gray-700">
-              <p className="text-xs text-gray-500 dark:text-gray-400 mb-1">Complexity</p>
-              <p className="text-lg font-semibold text-gray-900 dark:text-white">
-                {analysis.analysis?.complexity_score ?? detail.complexity_score ?? '-'}/10
-              </p>
-            </div>
-            <div className="bg-white dark:bg-gray-800 rounded-lg p-3 border border-gray-200 dark:border-gray-700">
-              <p className="text-xs text-gray-500 dark:text-gray-400 mb-1">Effort Level</p>
-              {(analysis.analysis?.effort_level || detail.effort_level) ? (
-                <span className={cn('px-2 py-1 rounded-full text-xs font-medium', effortColors[analysis.analysis?.effort_level || detail.effort_level || ''])}>
-                  {analysis.analysis?.effort_level || detail.effort_level}
-                </span>
-              ) : (
-                <span className="text-sm text-gray-400">-</span>
-              )}
-            </div>
-            <div className="bg-white dark:bg-gray-800 rounded-lg p-3 border border-gray-200 dark:border-gray-700">
-              <p className="text-xs text-gray-500 dark:text-gray-400 mb-1">Subtasks</p>
-              <p className="text-lg font-semibold text-gray-900 dark:text-white">
-                {analysis.subtask_count ?? detail.subtask_count ?? '-'}
-              </p>
-            </div>
-            <div className="bg-white dark:bg-gray-800 rounded-lg p-3 border border-gray-200 dark:border-gray-700">
-              <p className="text-xs text-gray-500 dark:text-gray-400 mb-1">Strategy</p>
-              <div className="flex items-center gap-1">
-                {(() => {
-                  const strategy = analysis.strategy || detail.strategy || ''
-                  const StrategyIcon = strategyIcons[strategy] || GitBranch
-                  return <StrategyIcon className="w-4 h-4 text-primary-500" />
-                })()}
-                <span className="text-sm font-medium text-gray-900 dark:text-white capitalize">
-                  {analysis.strategy || detail.strategy || '-'}
-                </span>
-              </div>
-            </div>
-          </div>
-
-          {/* Context Summary */}
-          {analysis.analysis?.context_summary && (
-            <div className="bg-white dark:bg-gray-800 rounded-lg p-3 border border-gray-200 dark:border-gray-700">
-              <h4 className="text-xs font-medium text-gray-500 dark:text-gray-400 mb-1.5">Context Summary</h4>
-              <p className="text-sm text-gray-600 dark:text-gray-400">
-                {analysis.analysis.context_summary}
-              </p>
-            </div>
-          )}
-
-          {/* Key Requirements */}
-          {analysis.analysis?.key_requirements && analysis.analysis.key_requirements.length > 0 && (
-            <div>
-              <h4 className="text-xs font-medium text-gray-500 dark:text-gray-400 mb-2">Key Requirements</h4>
-              <div className="flex flex-wrap gap-1.5">
-                {analysis.analysis.key_requirements.map((req, i) => (
-                  <span
-                    key={i}
-                    className="px-2 py-0.5 bg-primary-100 dark:bg-primary-900/30 text-primary-700 dark:text-primary-400 text-xs rounded-full"
-                  >
-                    {req}
-                  </span>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {/* Execution Plan */}
-          {analysis.execution_plan?.parallel_groups && analysis.execution_plan.parallel_groups.length > 0 && (
-            <div>
-              <h4 className="text-xs font-medium text-gray-500 dark:text-gray-400 mb-2">Execution Plan</h4>
-              <div className="space-y-2">
-                {analysis.execution_plan.parallel_groups.map((group, groupIndex) => (
-                  <div key={groupIndex}>
-                    <div className="flex items-center gap-2 mb-1.5">
-                      <span className="text-xs font-medium text-gray-500 dark:text-gray-400">
-                        Step {groupIndex + 1}
-                      </span>
-                      {group.length > 1 && (
-                        <span className="px-1.5 py-0.5 bg-blue-100 dark:bg-blue-900/30 text-blue-700 dark:text-blue-400 text-[10px] rounded-full flex items-center gap-0.5">
-                          <Zap className="w-2.5 h-2.5" />
-                          Parallel
-                        </span>
-                      )}
-                    </div>
-                    <div className={cn('grid gap-2', group.length > 1 ? 'grid-cols-1 md:grid-cols-2' : 'grid-cols-1')}>
-                      {group.map((taskId) => {
-                        const subtask = analysis.execution_plan?.subtasks[taskId]
-                        if (!subtask) return null
-                        return (
-                          <div
-                            key={taskId}
-                            className="bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 rounded-lg p-2.5"
-                          >
-                            <div className="flex items-start justify-between mb-1">
-                              <h5 className="text-xs font-medium text-gray-900 dark:text-white">
-                                {subtask.title}
-                              </h5>
-                              <span className={cn('px-1.5 py-0.5 text-[10px] rounded-full flex-shrink-0 ml-2', effortColors[subtask.effort])}>
-                                {subtask.effort}
-                              </span>
-                            </div>
-                            {subtask.agent && (
-                              <div className="flex items-center gap-1 text-[10px] text-gray-500 dark:text-gray-400">
-                                <Users className="w-2.5 h-2.5" />
-                                <span>{subtask.agent}</span>
-                              </div>
-                            )}
-                          </div>
-                        )
-                      })}
-                    </div>
-                    {groupIndex < analysis.execution_plan!.parallel_groups.length - 1 && (
-                      <div className="flex justify-center my-1">
-                        <ChevronRight className="w-4 h-4 text-gray-300 dark:text-gray-600 rotate-90" />
-                      </div>
-                    )}
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-        </>
-      )}
-
-      {/* Fallback for no analysis data */}
-      {!analysis && (
-        <div className="text-sm text-gray-500 dark:text-gray-400">
-          <p>분석 결과: {detail.success ? '성공' : '실패'}</p>
-          {detail.error && <p className="text-red-500 mt-1">{detail.error}</p>}
-          {detail.execution_time_ms > 0 && (
-            <p className="mt-1">실행 시간: {detail.execution_time_ms}ms</p>
-          )}
-        </div>
-      )}
-    </div>
-  )
-}
-
-function ActivityHeatmapChart({ data }: { data: ActivityHeatmap }) {
-  const cellHeight = 16
-  const isDark = document.documentElement.classList.contains('dark')
-  const emptyColor = isDark ? '#171f2a' : '#e5e7eb'
-
-  return (
-    <div className="w-full overflow-hidden">
-      <div className="flex gap-1 w-full">
-        {/* Hour labels */}
-        <div className="flex flex-col gap-[2px] text-[10px] text-gray-500 pr-1 flex-shrink-0 w-8">
-          {Array.from({ length: 24 }, (_, i) => (
-            <div
-              key={i}
-              style={{ height: cellHeight }}
-              className="flex items-center justify-end leading-none"
-            >
-              {i % 6 === 0 ? `${i}:00` : ''}
-            </div>
-          ))}
-        </div>
-
-        {/* Heatmap grid */}
-        {Array.from({ length: 7 }, (_, day) => (
-          <div key={day} className="flex flex-col gap-[2px] flex-1 min-w-0">
-            {Array.from({ length: 24 }, (_, hour) => {
-              const cell = data.cells.find((c) => c.day === day && c.hour === hour)
-              const value = cell?.value || 0
-              const alpha = computeHeatmapAlpha(value, data.max_value)
-
-              return (
-                <div
-                  key={hour}
-                  style={{
-                    height: cellHeight,
-                    backgroundColor: alpha === 0
-                      ? emptyColor
-                      : `rgba(59, 130, 246, ${alpha})`,
-                  }}
-                  className="w-full rounded-sm cursor-pointer hover:ring-2 hover:ring-blue-400"
-                  title={`${DAY_LABELS[day]} ${hour}:00 - ${value} sessions`}
-                />
-              )
-            })}
-            <div className="text-[10px] text-gray-500 text-center mt-1 truncate">
-              {DAY_LABELS[day]}
-            </div>
-          </div>
-        ))}
-      </div>
-
-      {/* Legend */}
-      <div className="flex items-center gap-2 mt-4 text-xs text-gray-500">
-        <span>Less</span>
-        {[0, 0.25, 0.5, 0.75, 1].map((intensity, i) => (
-          <div
-            key={i}
-            style={{
-              width: 12,
-              height: 12,
-              backgroundColor: intensity === 0
-                ? emptyColor
-                : `rgba(59, 130, 246, ${0.2 + intensity * 0.8})`,
-            }}
-            className="rounded-sm"
-          />
-        ))}
-        <span>More</span>
-      </div>
-    </div>
-  )
-}
-
-// ─────────────────────────────────────────────────────────────
-// Utilities
-// ─────────────────────────────────────────────────────────────
-
-function formatNumber(num: number): string {
-  if (num >= 1_000_000) return `${(num / 1_000_000).toFixed(1)}M`
-  if (num >= 1_000) return `${(num / 1_000).toFixed(1)}K`
-  return num.toString()
-}
-
-function formatTokenCount(value: number): string {
-  if (value >= 1_000_000) return `${(value / 1_000_000).toFixed(1)}M`
-  if (value >= 1_000) return `${(value / 1_000).toFixed(1)}K`
-  return String(value)
-}
-
-function formatDuration(ms: number): string {
-  if (ms < 1000) return `${ms}ms`
-  if (ms < 60000) return `${(ms / 1000).toFixed(1)}s`
-  return `${(ms / 60000).toFixed(1)}m`
-}
-
-function formatTrendLabel(timestamp: string, timeRange: TimeRange): string {
-  const date = new Date(timestamp)
-  switch (timeRange) {
-    case '1h':
-      return date.toLocaleTimeString('ko-KR', { hour: '2-digit', minute: '2-digit' })
-    case '24h':
-      return date.toLocaleTimeString('ko-KR', { hour: '2-digit', minute: '2-digit' })
-    case '7d':
-      return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' })
-    case '30d':
-      return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' })
-    case 'all':
-      return date.toLocaleDateString('en-US', { year: '2-digit', month: 'short' })
-    default:
-      return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' })
-  }
-}
-
-function formatTrendData(
-  data: TrendDataPoint[],
-  timeRange: TimeRange = '7d',
-): { label: string; value: number | undefined }[] {
-  return data.map((d) => ({
-    label: formatTrendLabel(d.timestamp, timeRange),
-    value: d.value ?? undefined, // null → undefined for Recharts gap handling
-  }))
-}
-
-/**
- * Transform multi-project series data for Recharts
- * Converts from series-based to date-based format
- */
-function transformMultiSeriesData(
-  response: MultiProjectTrendsResponse
-): Record<string, string | number | null>[] {
-  const dateMap = new Map<string, Record<string, string | number | null>>()
-
-  response.series.forEach((series) => {
-    series.data.forEach((point) => {
-      const dateKey = formatTrendLabel(point.timestamp, response.period)
-      const existing = dateMap.get(dateKey) || { date: dateKey }
-      existing[series.project_id] = point.value
-      dateMap.set(dateKey, existing)
-    })
-  })
-
-  return Array.from(dateMap.values())
-}
-
-function buildModelTokenBreakdown(
-  models: CostBreakdown[],
-  externalSummary: ExternalUsageSummaryResponse | null,
-): ModelTokenBreakdown[] {
-  const grouped = new Map<string, ModelTokenBreakdown>()
-
-  const addEntry = (
-    modelName: string | null | undefined,
-    providerValue: string | null | undefined,
-    tokens: number,
-    cost: number,
-  ) => {
-    if (tokens <= 0) return
-    const model = modelName?.trim() || 'unknown'
-    const provider = normalizeProvider(providerValue, model)
-    const key = `${provider}:${model}`
-    const existing = grouped.get(key)
-
-    if (existing) {
-      existing.tokens += tokens
-      existing.cost += cost
-      return
-    }
-
-    grouped.set(key, {
-      model,
-      provider,
-      providerLabel: PROVIDER_LABELS[provider] ?? provider,
-      tokens,
-      cost,
-      percentage: 0,
-      color: PROVIDER_COLORS[provider] ?? PROVIDER_COLORS.unknown,
-    })
-  }
-
-  models.forEach((model) => {
-    addEntry(model.value, model.provider, model.tokens, model.cost)
-  })
-
-  externalSummary?.records.forEach((record) => {
-    addEntry(record.model, record.provider, record.total_tokens, record.cost_usd)
-  })
-
-  const entries = Array.from(grouped.values())
-  const totalTokens = entries.reduce((sum, entry) => sum + entry.tokens, 0)
-  return entries
-    .map((entry) => ({
-      ...entry,
-      percentage: totalTokens > 0 ? Math.round((entry.tokens / totalTokens) * 1000) / 10 : 0,
-    }))
-    .sort((a, b) => b.tokens - a.tokens)
-}
-
-function filterAttributedModelPerformance(agents: AgentPerformance[]): AgentPerformance[] {
-  return agents.filter((agent) => {
-    const modelName = agent.agent_name.trim().toLowerCase()
-    return !(
-      modelName === 'unknown (no model info)'
-      && agent.total_tokens === 0
-      && agent.total_cost === 0
-    )
-  })
-}
-
-function normalizeProvider(provider: string | null | undefined, modelName: string): string {
-  const explicit = provider?.trim().toLowerCase()
-  if (explicit) {
-    if (explicit === 'codex') return 'codex_cli'
-    if (explicit === 'claude') return 'anthropic'
-    if (explicit === 'gemini' || explicit === 'vertex' || explicit === 'google_gemini') {
-      return 'google'
-    }
-    return explicit
-  }
-
-  const model = modelName.toLowerCase()
-  if (model.includes('codex')) return 'codex_cli'
-  if (model.includes('claude')) return 'anthropic'
-  if (model.includes('gpt') || model.includes('openai')) return 'openai'
-  if (model.includes('gemini') || model.includes('vertex')) return 'google'
-  if (model.includes('ollama') || model.includes('llama') || model.includes('mistral')) {
-    return 'ollama'
-  }
-  return 'unknown'
-}
-
-function renderAosModelSourceBadge(models: ModelTokenBreakdown[]): React.ReactNode {
-  if (models.length === 0) return null
-  const providers = Array.from(new Set(models.map((model) => model.providerLabel)))
-  const label = providers.length === 1 ? providers[0] : `${providers.length} providers`
-
-  return (
-    <span
-      className="text-[11px] px-2 py-0.5 rounded-full bg-blue-100 text-blue-800 dark:bg-blue-900/40 dark:text-blue-200 whitespace-nowrap"
-      title="AOS analytics costs.by_model 기준의 모델별 LLM 사용량입니다."
-    >
-      AOS 집계: {label}
-    </span>
-  )
-}
-
-function truncateModelLabel(modelName: string): string {
-  return modelName.length > 18 ? `${modelName.slice(0, 17)}...` : modelName
 }
 
 export default AnalyticsPage
